form input logic moved to component
