+++ src/components/ui/index.js
@@ -2,6 +2,7 @@ import Button from './button'
 import Card from './card'
 import Checkbox from './checkbox'
 import FormInput from './form-input'
+import FormInputWrap from './form-input-wrap'
 import Modal from './modal'
 
 export {
@@ -9,5 +10,6 @@ export {
   Card,
   Checkbox,
   FormInput,
+  FormInputWrap,
   Modal
 }
